feat(parser): add `parse_parameters` function
Introduce function to parse the parameters section
of a YAML specification file.
The function enables processing parameter sets
with support for `factory` resolution
and generic value handling.

Refs: #452, #462
Closes: #472
